+++ lib/sync/import.dart
@@ -9,10 +9,12 @@ import "package:flow/entity/category.dart";
 import "package:flow/entity/transaction.dart";
 import "package:flow/sync/exception.dart";
 import "package:flow/sync/import/base.dart";
+import "package:flow/sync/import/external/ivy_wallet_csv.dart";
 import "package:flow/sync/import/import_csv.dart";
 import "package:flow/sync/import/import_v1.dart";
 import "package:flow/sync/import/import_v2.dart";
 import "package:flow/sync/model/csv/parsed_data.dart";
+import "package:flow/sync/model/external/ivy/ivy_wallet_csv.dart";
 import "package:flow/sync/model/model_v1.dart";
 import "package:flow/sync/model/model_v2.dart";
 import "package:flow/utils/utils.dart";
@@ -22,6 +24,14 @@ import "package:path_provider/path_provider.dart";
 export "package:flow/sync/import/import_v1.dart";
 export "package:flow/sync/model/model_v1.dart";
 
+enum ImportExternalFormat {
+  ivyWallet("Ivy Wallet");
+
+  final String name;
+
+  const ImportExternalFormat(this.name);
+}
+
 /// We have to recover following models:
 /// * Account
 /// * Category
@@ -30,14 +40,33 @@ export "package:flow/sync/model/model_v1.dart";
 ///
 /// We need to resolve [Transaction]s last cause it references both [Account] and
 /// [Category] UUID.
-Future<Importer> importBackup({File? backupFile}) async {
-  final file = backupFile ?? await pickJsonOrZipFile();
+Future<Importer> importBackup({
+  File? backupFile,
+  ImportExternalFormat? externalFormat,
+}) async {
+  final file = backupFile ?? await pickImportFile();
 
   if (file == null) {
     throw const ImportException(
       "No file was picked to proceed with the import",
       l10nKey: "error.input.noFilePicked",
     );
+  }
+
+  if (externalFormat == null) {
+    final String baseName = path.basename(file.path);
+
+    if (baseName.startsWith("Ivy Wallet")) {
+      externalFormat = ImportExternalFormat.ivyWallet;
+    }
+  }
+
+  if (externalFormat != null) {
+    return switch (externalFormat) {
+      ImportExternalFormat.ivyWallet => IvyWalletCsvImporter(
+        await IvyWalletCsv.fromFile(file),
+      ),
+    };
   }
 
   final String ext = path.extension(file.path).toLowerCase();